requirements.txt
mcp>=1.0.0
mysql-connector-python>=9.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Vector database and AI dependencies
pymongo>=4.6.0
motor>=3.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
